# golang/docs/swagger.yaml
basePath: /api
definitions:
...
  api.Prompt:
    properties:
      id:
        type: string
      text:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a API documentation for OPENAI.
  title: OPENAI
  version: "1.0"
paths:
...
  /v1/prompt/{id}:
    get:
      description: Retrieves a prompt by ID
      parameters:
      - description: Prompt ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/api.Prompt'
      summary: Retrieve a prompt by ID
      tags:
      - v1
  /v2/prompt/{id}:
    get:
      description: Retrieves a prompt by ID
      parameters:
      - description: Prompt ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/api.Prompt'
      summary: Retrieve a prompt by ID
      tags:
      - v2
swagger: "2.0"